% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launch_app}
\alias{launch_app}
\title{Launch oglcnac Shiny App}
\usage{
launch_app()
}
\value{
None
}
\description{
This function launches a Shiny App for uploading, processing,
and downloading UniProt data in CSV, TSV, or Excel format.
Users can upload data, preview it, and select specific columns for processing.
The processed data can be viewed and downloaded.
}
\examples{
if (interactive()) {
  oglcnac::launch_app()
}
}
