% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_uniprot_data}
\alias{retrieve_uniprot_data}
\title{Retrieve Data from UniProt API}
\usage{
retrieve_uniprot_data(accession)
}
\arguments{
\item{accession}{A character string representing the UniProt accession number.}
}
\value{
A list containing the retrieved data in JSON format, or NULL if the request fails.
}
\description{
This function sends a GET request to the UniProt REST API and retrieves
data based on the provided UniProt accession number.
}
\examples{
# Example usage
result <- retrieve_uniprot_data("O88737")
print(result)
}
